Moved similar marker related code to super class
- handle markers on paths with successive equal points
 (fixing drawing of __issue-345-01.svg)
