+++ Source/Basic Shapes/SvgLine.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Drawing;
 using System.Drawing.Drawing2D;
-using Svg.ExtensionMethods;
 
 namespace Svg
 {
@@ -9,7 +7,7 @@ namespace Svg
     /// Represents and SVG line element.
     /// </summary>
     [SvgElement("line")]
-    public class SvgLine : SvgPathBasedElement
+    public class SvgLine : SvgMarkerElement
     {
         private SvgUnit _startX;
         private SvgUnit _startY;
@@ -77,38 +75,6 @@ namespace Svg
             }
         }
 
-        /// <summary>
-        /// Gets or sets the marker (end cap) of the path.
-        /// </summary>
-        [SvgAttribute("marker-end")]
-        public Uri MarkerEnd
-        {
-            get { return this.Attributes.GetAttribute<Uri>("marker-end").ReplaceWithNullIfNone(); }
-            set { this.Attributes["marker-end"] = value; }
-        }
-
-
-        /// <summary>
-        /// Gets or sets the marker (start cap) of the path.
-        /// </summary>
-        [SvgAttribute("marker-mid")]
-        public Uri MarkerMid
-        {
-            get { return this.Attributes.GetAttribute<Uri>("marker-mid").ReplaceWithNullIfNone(); }
-            set { this.Attributes["marker-mid"] = value; }
-        }
-
-
-        /// <summary>
-        /// Gets or sets the marker (start cap) of the path.
-        /// </summary>
-        [SvgAttribute("marker-start")]
-        public Uri MarkerStart
-        {
-	        get { return this.Attributes.GetAttribute<Uri>("marker-start").ReplaceWithNullIfNone(); }
-            set { this.Attributes["marker-start"] = value; }
-        }
-
         public override SvgPaintServer Fill
         {
             get { return null; /* Line can't have a fill */ }
@@ -150,37 +116,6 @@ namespace Svg
                 }
             }
             return this._path;
-        }
-
-        /// <summary>
-        /// Renders the stroke of the <see cref="SvgVisualElement"/> to the specified <see cref="ISvgRenderer"/>
-        /// </summary>
-        /// <param name="renderer">The <see cref="ISvgRenderer"/> object to render to.</param>
-        protected internal override bool RenderStroke(ISvgRenderer renderer)
-        {
-            var result = base.RenderStroke(renderer);
-            var path = this.Path(renderer);
-
-            if (this.MarkerStart != null)
-            {
-                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerStart.ToString());
-                marker.RenderMarker(renderer, this, path.PathPoints[0], path.PathPoints[0], path.PathPoints[1]);
-            }
-
-            if (this.MarkerMid != null)
-            {
-                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerMid.ToString());
-                for (int i = 1; i <= path.PathPoints.Length - 2; i++)
-                    marker.RenderMarker(renderer, this, path.PathPoints[i], path.PathPoints[i - 1], path.PathPoints[i], path.PathPoints[i + 1]);
-            }
-
-            if (this.MarkerEnd != null)
-            {
-                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerEnd.ToString());
-                marker.RenderMarker(renderer, this, path.PathPoints[path.PathPoints.Length - 1], path.PathPoints[path.PathPoints.Length - 2], path.PathPoints[path.PathPoints.Length - 1]);
-            }
-
-            return result;
         }
 
 		public override SvgElement DeepCopy()
