+++ Source/Basic Shapes/SvgPolygon.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Drawing.Drawing2D;
 using System.Diagnostics;
-using Svg.ExtensionMethods;
 
 namespace Svg
 {
@@ -9,7 +7,7 @@ namespace Svg
     /// SvgPolygon defines a closed shape consisting of a set of connected straight line segments.
     /// </summary>
     [SvgElement("polygon")]
-    public class SvgPolygon : SvgPathBasedElement
+    public class SvgPolygon : SvgMarkerElement
     {
         private GraphicsPath _path;
         
@@ -21,38 +19,6 @@ namespace Svg
         {
             get { return this.Attributes["points"] as SvgPointCollection; }
             set { this.Attributes["points"] = value; this.IsPathDirty = true; }
-        }
-
-        /// <summary>
-        /// Gets or sets the marker (end cap) of the path.
-        /// </summary>
-        [SvgAttribute("marker-end")]
-        public virtual Uri MarkerEnd
-        {
-            get { return this.Attributes.GetAttribute<Uri>("marker-end").ReplaceWithNullIfNone(); }
-            set { this.Attributes["marker-end"] = value; }
-        }
-
-
-        /// <summary>
-        /// Gets or sets the marker (start cap) of the path.
-        /// </summary>
-        [SvgAttribute("marker-mid")]
-        public virtual Uri MarkerMid
-        {
-            get { return this.Attributes.GetAttribute<Uri>("marker-mid").ReplaceWithNullIfNone(); }
-            set { this.Attributes["marker-mid"] = value; }
-        }
-
-
-        /// <summary>
-        /// Gets or sets the marker (start cap) of the path.
-        /// </summary>
-        [SvgAttribute("marker-start")]
-        public virtual Uri MarkerStart
-        {
-            get { return this.Attributes.GetAttribute<Uri>("marker-start").ReplaceWithNullIfNone(); }
-            set { this.Attributes["marker-start"] = value; }
         }
 
         public override GraphicsPath Path(ISvgRenderer renderer)
@@ -99,37 +65,6 @@ namespace Svg
                   this.IsPathDirty = false;
             }
             return this._path;
-        }
-
-        /// <summary>
-        /// Renders the stroke of the <see cref="SvgVisualElement"/> to the specified <see cref="ISvgRenderer"/>
-        /// </summary>
-        /// <param name="renderer">The <see cref="ISvgRenderer"/> object to render to.</param>
-        protected internal override bool RenderStroke(ISvgRenderer renderer)
-        {
-            var result = base.RenderStroke(renderer);
-            var path = this.Path(renderer);
-
-            if (this.MarkerStart != null)
-            {
-                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerStart.ToString());
-                marker.RenderMarker(renderer, this, path.PathPoints[0], path.PathPoints[0], path.PathPoints[1]);
-            }
-
-            if (this.MarkerMid != null)
-            {
-                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerMid.ToString());
-                for (int i = 1; i <= path.PathPoints.Length - 2; i++)
-                    marker.RenderMarker(renderer, this, path.PathPoints[i], path.PathPoints[i - 1], path.PathPoints[i], path.PathPoints[i + 1]);
-            }
-
-            if (this.MarkerEnd != null)
-            {
-                SvgMarker marker = this.OwnerDocument.GetElementById<SvgMarker>(this.MarkerEnd.ToString());
-                marker.RenderMarker(renderer, this, path.PathPoints[path.PathPoints.Length - 1], path.PathPoints[path.PathPoints.Length - 2], path.PathPoints[path.PathPoints.Length - 1]);
-            }
-
-            return result;
         }
 
 		public override SvgElement DeepCopy()
