+++ Source/Svg.csproj
@@ -91,6 +91,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Basic Shapes\SvgImage.cs" />
+    <Compile Include="Basic Shapes\SvgMarkerElement.cs" />
     <Compile Include="Basic Shapes\SvgPathBasedElement.cs" />
     <Compile Include="Basic Shapes\SvgVisualElement.cs" />
     <Compile Include="Basic Shapes\SvgCircle.cs" />
