+++ Tests/Svg.UnitTests/PassingTests.csv
@@ -187,6 +187,7 @@ __issue-323-03
 __issue-329-01
 __issue-338-01_stroke_width
 __issue-342-01
+__issue-345-01
 __issue-354-01
 __issue-385-01_Test_text-anchor-middle
 __issue-398-01
